Create MATLAB CODE
Two-joint planar robot arm: If L1 and L2 are known then specifying θ1 and θ2 will determine all the points on the robot.Thus, configuration is q =(θ1 , θ2 ) Each 𝜃𝜃i correspond to a point on a unit circle 𝑆𝑆1 So, Configuration space is 𝑆𝑆1 × 𝑆𝑆1 = 𝑇𝑇1 - 2D torusThe configuration space plotted in MATLAB as 2d plane ,can be visualized as 2D torus by folding the plane such that θ1=0 omeet with θ1 =360 o(horizontal folding)and similarly for θ2 i.e folding vertically such that θ2=0omeet with θ2 =360 o  because after 360 o ,angle just repeats as after 0 o. 
(360 o+θ= θ),which can be visualized in 2nd plot of Workspace, depicting the robot end effector workspace.*For MATLAB coding we have use some offset w.r.t to both obstacles boundary, so theresponse is good (Conservative Approach),as we have used discreate angle with a step of 10oand if it collide with Obstacles, it can be plotted in C-Space correctly as obstacles configuration space with corresponding colour of obstacles.In this ,Obstacles created and filled with different colours to depict them easily in C-Space and to get Real World Animation, I have used the set handle command to get the arms at correct position of theta1 and theta2 using the rotation transformation matrix and giving an translation to (x1,y1) and (x2,y2) to get (X1,Y2) and (X2,Y2) for proper viewing ,so 1st arm is appended to origin and the 2nd arm is appended correctly to 1st one (as given is Cpace.m file) and then the end effector is being plotted using scatter command and when the two-joint planar robot arm overlap/collide with any obstacles ,it plot the corresponding theta1 and theta 2 in Cspace as Obstacles Configuration space with corresponding obstacles color, keeping the free space as white/no colour.Because of step angle used (10o),the workspace has some gaps in-between, but it should be continuous i.e. subset of 2-d Real Space(R2). One more Observation is that, viewing the robot in real space and finding the path can be hectic, but in C-Space ,we just need to find the connected path between start position and goal position, avoiding the obstacles Configuration space. In C-Space, any two-point connected, even if they are neighbour, there lie many angles in between, which may not be connected. So, it’s not feasible to find a path in C-Space as a straight line, because any two neighbouring pixels in C-Space is actually more curved rather then connected by a straight line.So, C-Space is basically a Conceptual tool and it’s help us in finding a path in a space, where Robot Configuration and Obstacles Configuration is very easy to visualize, but practically it’s not feasible because of computational complexity and also to find every possibility of configuration it’s not practical.